New Update
Linked in Login
Partially functioning settings page.
Auto card creation on login.
boolean fields on settings tied into login for social networks.
still need to add twitter and instagram, need to merge this build with
current or apply bug fixes to this build

+++ REON/SmallMonthCell.m
@@ -49,7 +49,7 @@
     //--- Event Cells
     else if(cellYear <= currentYear && cellMonth <= currentMonth){
         
-        NSString *keyString = [NSString stringWithFormat:@"%li%li%li", cellMonth, (indexPath.row+1), cellYear];
+        NSString *keyString = [NSString stringWithFormat:@"%li%i%li", (long)cellMonth, (indexPath.row+1), (long)cellYear];
         
         if([eventDictionary objectForKey:keyString]){
             cell.isEvent = YES;
